src/requirements.txt
apscheduler==3.10.4 
python-dotenv==1.0.0
selenium  
webdriver-manager  
pymysql
pandas